Moved headers to appropriate places. Added overloads that accept the orientation of the input polygon.
