+++ Set_movable_separability_2/doc/Set_movable_separability_2/CGAL/Set_movable_separability_2/Single_mold_translational_casting/pullout_directions.h
@@ -2,7 +2,7 @@ namespace CGAL {
 namespace Set_movable_separability_2 {
 namespace Single_mold_translational_casting {
 
-/*! \ingroup PkgSetMovableSeparability2Funcs
+/*! \ingroup pullout_directions_grp
  *
  * Given a simple polygon and an edge of the polygon, this function determines
  * whether a cavity (of a mold in the plane) that has the shape of the polygon
@@ -15,9 +15,9 @@ namespace Single_mold_translational_casting {
  * The type that substitutes the template parameter `%CastingTraits_2` must be
  * a model of the concept `CastingTraits_2`.
  *
- * \param[in] pgn the input polygon.
- * \param[in] it an iterator to an edge in pgn.
- * \param[in] traits the traits to use.
+ * \param polygon the input polygon.
+ * \param it an iterator to an edge in polygon.
+ * \param traits the traits to use.
  * \return a pair of elements, where the first is a Boolean that indicates
  *         whether the input edge is a valid top edge, and the second
  *         is a closed range of pullout directions represented as a pair
@@ -31,21 +31,21 @@ template <typename CastingTraits_2>
 std::pair<bool, std::pair<typename CastingTraits_2::Direction_2,
                           typename CastingTraits_2::Direction_2> >
 pullout_directions
-(const CGAL::Polygon_2<CastingTraits_2>& pgn,
+(const CGAL::Polygon_2<CastingTraits_2>& polygon,
  const typename CGAL::Polygon_2<CastingTraits_2>::Edge_const_iterator& it,
  CastingTraits_2& traits = CastingTraits_2());
 
-/*! \ingroup PkgSetMovableSeparability2Funcs
+/*! \ingroup pullout_directions_grp
  *
  * Same as above with the additional `orientation` argument.
  * If the orientation of the polygon is known upon invocation, specify it.
  * Otherwise, it has to be computed.  Note that finding the orientation of a
  * polygon requires time linear in the number of edges.
  *
- * \param[in] pgn the input polygon.
- * \param[in] it an iterator to an edge in pgn.
- * \param[in] orientation the orientation of `pgn`.
- * \param[in] traits the traits to use.
+ * \param polygon the input polygon.
+ * \param it an iterator to an edge in polygon.
+ * \param orientation the orientation of `polygon`.
+ * \param traits the traits to use.
  * \return a pair of elements, where the first is a Boolean that indicates
  *         whether the input edge is a valid top edge, and the second
  *         is a closed range of pullout directions represented as a pair
@@ -59,7 +59,7 @@ template <typename CastingTraits_2>
 std::pair<bool, std::pair<typename CastingTraits_2::Direction_2,
                           typename CastingTraits_2::Direction_2> >
 pullout_directions
-(const CGAL::Polygon_2<CastingTraits_2>& pgn,
+(const CGAL::Polygon_2<CastingTraits_2>& polygon,
  const typename CGAL::Polygon_2<CastingTraits_2>::Edge_const_iterator& it,
  CGAL::Orientation orientation,
  CastingTraits_2& traits = CastingTraits_2());
