+++ Set_movable_separability_2/examples/Set_movable_separability_2/is_pullout_direction_single_mold_translational_casting.cpp
@@ -1,7 +1,7 @@
 #include <fstream>
 
 #include <CGAL/Exact_predicates_exact_constructions_kernel.h>
-#include <CGAL/is_pullout_direction_single_mold_translational_casting_2.h>
+#include <CGAL/Set_movable_separability_2/Single_mold_translational_casting/is_pullout_direction.h>
 
 typedef CGAL::Exact_predicates_exact_constructions_kernel Kernel;
 typedef Kernel::Vector_2                              	  Vector_2;
