+++ Set_movable_separability_2/test/Set_movable_separability_2/cgal_test_with_cmake
@@ -96,11 +96,14 @@ if [ $# -ne 0 ] ; then
   done
 else
   echo "Run all tests."
-if  ${MAKE_CMD} -f Makefile help | grep -E "test_single_mold_translational_casting$" > /dev/null; then
-  compile_and_run test_single_mold_translational_casting
-  compile_and_run test_is_pullout_directions_single_mold_translational
-  NEED_CLEAN=y
-fi
+  if  ${MAKE_CMD} -f Makefile help | grep -E "test_top_edges_single_mold_translational_casting$" > /dev/null; then
+    compile_and_run test_top_edges_single_mold_translational_casting
+    NEED_CLEAN=y
+  fi
+  if  ${MAKE_CMD} -f Makefile help | grep -E "test_is_pullout_directions_single_mold_translational_casting$" > /dev/null; then
+    compile_and_run test_is_pullout_directions_single_mold_translational_casting
+    NEED_CLEAN=y
+  fi
 fi
 
 #
